Remove server i18n HOC and refactor name of client one
