+++ pages/benefits-directory.js
@@ -1,7 +1,7 @@
 import React, { Component } from "react";
 import Router from "next/router";
 import PropTypes from "prop-types";
-import withI18N from "../lib/i18nClient";
+import withI18N from "../lib/i18nHOC";
 import Layout from "../components/layout";
 import { connect } from "react-redux";
 import BB from "../components/BB";
